TypeName

+++ modules/task_2/elandaev_p_integral_rectangle_omp/rectangle.cpp
@@ -87,7 +87,7 @@ double RecIntOmp(std::vector<double> start,
         throw "wrong splitting: smallest";
     std::vector<std::vector<double> > point;
     point.resize(start.size());
-    long long countTrial = 1;
+    int countTrial = 1;
     for (int i = 0; i < size; i++) {
         for (int j = 0; j < countstep[i]; j++) {
             point[i].push_back(start[i] + j * step);
@@ -140,7 +140,7 @@ double RecIntT(std::vector<double> start,
         throw "wrong splitting: smallest";
     std::vector<std::vector<double> > point;
     point.resize(start.size());
-    long long countTrial = 1;
+    int countTrial = 1;
     for (int i = 0; i < size; i++) {
         for (int j = 0; j < countstep[i]; j++) {
             point[i].push_back(start[i] + j * step);
